Refactoring: Moved `load_kube_config()` in `api.kubernetes`
where it belongs
